+++ NEWS.md
@@ -1,5 +1,6 @@
 # Wavelength 2.3.6
-* Resolve bug in Q3 where hierarchy in DATIM was adjusted 
+* Resolve bug in Q3 where hierarchy in DATIM was adjusted
+* Update `iso_map` due to new hierarchy
 
 # Wavelength 2.3.3
 * Resolve bug that converted NAs to zeros when wide formats were used